// src/permission/enum/permission.enum.ts
export enum PermissionEnum {
  // User Permissions
  CREATE_USER = 'CREATE_USER',
  UPDATE_USER = 'UPDATE_USER',
  VIEW_USER = 'VIEW_USER',
  DELETE_USER = 'DELETE_USER',

  // Issue Permissions
  CREATE_ISSUE = 'CREATE_ISSUE',
  UPDATE_ISSUE = 'UPDATE_ISSUE',
  VIEW_ISSUE = 'VIEW_ISSUE',
  DELETE_ISSUE = 'DELETE_ISSUE',
  UPDATE_ISSUE_STATUS = 'UPDATE_ISSUE_STATUS',
  UPDATE_ISSUE_PRIORITY = 'UPDATE_ISSUE_PRIORITY',
  UPDATE_ISSUE_SEVERITY = 'UPDATE_ISSUE_SEVERITY',

  // Role Permissions
  CREATE_ROLE = 'CREATE_ROLE',
  UPDATE_ROLE = 'UPDATE_ROLE',
  VIEW_ROLE = 'VIEW_ROLE',
  DELETE_ROLE = 'DELETE_ROLE',

  // Permission Management
  CREATE_PERMISSION = 'CREATE_PERMISSION',
  UPDATE_PERMISSION = 'UPDATE_PERMISSION',
  VIEW_PERMISSION = 'VIEW_PERMISSION',
  DELETE_PERMISSION = 'DELETE_PERMISSION',

  // Role and Permission Assignment
  ASSIGN_ROLE_PERMISSION = 'ASSIGN_ROLE_PERMISSION',
}